requirements.txt
together>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
tkinter-modern>=1.0.0
speechrecognition>=3.10.0
pyttsx3>=2.90
pyaudio>=0.2.11
playsound>=1.3.0